+++ src/invoices/InvoiceLinesSequence/index.js
@@ -1,0 +1,1 @@
+export { InvoiceLinesSequenceContainer as InvoiceLinesSequence } from './InvoiceLinesSequenceContainer';